+++ wxc_sdk/person_settings/common.py
@@ -81,6 +81,7 @@ class PersonSettingsApiChild(ApiChild, base=''):
             ('workspaces', 'outgoingPermission/digitPatterns'): ('telephony/config/workspaces', '/'),
             ('workspaces', 'privacy'): ('telephony/config/workspaces', '/'),
             ('workspaces', 'pushToTalk'): ('telephony/config/workspaces', '/'),
+            ('workspaces', 'selectiveReject'): ('telephony/config/workspaces', '/'),
             ('workspaces', 'sequentialRing'): ('telephony/config/workspaces', '/'),
             ('workspaces', 'simultaneousRing'): ('telephony/config/workspaces', '/'),
             ('workspaces', 'voicemail'): ('telephony/config/workspaces', '/'),