+++ wxc_sdk/person_settings/selective_reject.py
@@ -4,39 +4,18 @@ from wxc_sdk.base import ApiModel
 from wxc_sdk.common.selective import SelectiveCriteria, SelectiveCrit
 from wxc_sdk.person_settings.common import PersonSettingsApiChild
 
-__all__ = ['SequentialRingNumber', 'SequentialRing', 'SequentialRingApi', 'SequentialRingCriteria']
+__all__ = ['SelectiveRejectCriteria', 'SelectiveReject', 'SelectiveRejectApi']
 
 
-class SequentialRingCriteria(SelectiveCriteria):
-    _enabled_attr = 'ringEnabled'
+class SelectiveRejectCriteria(SelectiveCriteria):
+    _enabled_attr = 'rejectEnabled'
     _phone_numbers = 'phoneNumbers'
 
 
-class SequentialRingNumber(ApiModel):
-    #: Phone number set as the sequential number.
-    phone_number: Optional[str] = None
-    #: When set to `true` the called party is required to press 1 on the keypad to receive the call.
-    answer_confirmation_required_enabled: Optional[bool] = None
-    #: The number of rings to the specified phone number before the call advances to the subsequent number in the
-    #: sequence or goes to voicemail.
-    number_of_rings: Optional[int] = None
-
-
-class SequentialRing(ApiModel):
-    #: When set to `true` sequential ring is enabled.
+class SelectiveReject(ApiModel):
+    #: `true` if the Selective Reject feature is enabled.
     enabled: Optional[bool] = None
-    #: When set to `true`, the webex calling primary line will ring first.
-    ring_base_location_first_enabled: Optional[bool] = None
-    #: The number of times the primary line will ring.
-    base_location_number_of_rings: Optional[int] = None
-    #: When set to `true` and the primary line is busy, the system redirects calls to the numbers configured for
-    #: sequential ringing.
-    continue_if_base_location_is_busy_enabled: Optional[bool] = None
-    #: When set to `true` calls are directed to voicemail.
-    calls_to_voicemail_enabled: Optional[bool] = None
-    #: A list of up to five phone numbers to which calls will be directed.
-    phone_numbers: Optional[list[SequentialRingNumber]] = None
-    #: A list of criteria specifying conditions when sequential ringing is in effect.
+    #: A list of criteria specifying conditions when selective reject is in effect.
     criteria: Optional[list[SelectiveCrit]] = None
 
     def update(self) -> dict:
@@ -49,24 +28,25 @@ class SequentialRing(ApiModel):
                                exclude={'criteria'})
 
 
-class SequentialRingApi(PersonSettingsApiChild):
+class SelectiveRejectApi(PersonSettingsApiChild):
     """
-    API for sequential ring settings
+    API for selective reject settings
 
     For now only used for workspaces
     """
 
-    feature = 'sequentialRing'
+    feature = 'selectiveReject'
 
     def read_criteria(self, entity_id: str, id: str,
-                      org_id: str = None) -> SequentialRingCriteria:
+                      org_id: str = None) -> SelectiveRejectCriteria:
         """
-        Retrieve sequential ring criteria for an entity.
+        Retrieve Selective Reject Criteria for an entity
 
-        The sequential ring feature enables you to create a list of up to five phone numbers. When the workspace
-        receives incoming calls, these numbers will ring one after another.
-        The sequential ring criteria specify settings such as schedule and incoming numbers for which to sequentially
-        ring or not.
+        Retrieve Selective Reject Criteria Settings for an entity.
+
+        With the Selective Reject feature, you can reject calls at specific times from specific callers. This setting
+        takes precedence over Selectively Accept Calls.
+        Schedules can also be set up for this feature during certain times of the day or days of the week.
 
         **NOTE**: This API is only available for professional licensed workspaces.
 
@@ -78,25 +58,26 @@ class SequentialRingApi(PersonSettingsApiChild):
             organization (such as partners) may use this parameter as the default is the same organization as the
             token used to access API.
         :type org_id: str
-        :rtype: :class:`SelectiveCriteria`
+        :rtype: :class:`SelectiveRejectCriteria`
         """
         params = {}
         if org_id is not None:
             params['orgId'] = org_id
         url = self.f_ep(entity_id, f'criteria/{id}')
         data = super().get(url, params=params)
-        r = SequentialRingCriteria.model_validate(data)
+        r = SelectiveRejectCriteria.model_validate(data)
         return r
 
-    def configure_criteria(self, entity_id: str, id: str, settings: SequentialRingCriteria,
+    def configure_criteria(self, entity_id: str, id: str, settings: SelectiveRejectCriteria,
                            org_id: str = None):
         """
-        Modify sequential ring criteria for an entity.
+        Modify Selective Reject Criteria for an entity
 
-        The sequential ring feature enables you to create a list of up to five phone numbers. When the workspace
-        receives incoming calls, these numbers will ring one after another.
-        The sequential ring criteria specify settings such as schedule and incoming numbers for which to sequentially
-        ring or not.
+        Modify Selective Reject Criteria Settings for an entity.
+
+        With the Selective Reject feature, you can reject calls at specific times from specific callers. This setting
+        takes precedence over Selectively Accept Calls.
+        Schedules can also be set up for this feature during certain times of the day or days of the week.
 
         **NOTE**: This API is only available for professional licensed workspaces.
 
@@ -105,7 +86,7 @@ class SequentialRingApi(PersonSettingsApiChild):
         :param id: Unique identifier for the criteria.
         :type id: str
         :param settings: new settings to be applied.
-        :type settings: SequentialRingCriteria
+        :type settings: SelectiveRejectCriteria
         :param org_id: ID of the organization within which the entity resides. Only admin users of another
             organization (such as partners) may use this parameter as the default is the same organization as the
             token used to access API.
@@ -121,12 +102,13 @@ class SequentialRingApi(PersonSettingsApiChild):
 
     def delete_criteria(self, entity_id: str, id: str, org_id: str = None):
         """
-        Delete sequential ring criteria for an entity.
+        Delete Selective Reject Criteria for an entity
 
-        The sequential ring feature enables you to create a list of up to five phone numbers. When the workspace
-        receives incoming calls, these numbers will ring one after another.
-        The sequential ring criteria specify settings such as schedule and incoming numbers for which to sequentially
-        ring or not.
+        Delete Selective Reject criteria Settings for an entity.
+
+        With the Selective Reject feature, you can reject calls at specific times from specific callers. This setting
+        takes precedence over Selectively Accept Calls.
+        Schedules can also be set up for this feature during certain times of the day or days of the week.
 
         **NOTE**: This API is only available for professional licensed workspaces.
 
@@ -146,22 +128,22 @@ class SequentialRingApi(PersonSettingsApiChild):
         url = self.f_ep(entity_id, f'criteria/{id}')
         super().delete(url, params=params)
 
-    def create_criteria(self, entity_id: str, settings: SequentialRingCriteria,
-                        org_id: str = None) -> str:
+    def create_criteria(self, entity_id: str, settings: SelectiveRejectCriteria, org_id: str = None) -> str:
         """
-        Create sequential ring criteria for an entity.
+        Create Selective Reject Criteria for an entity
 
-        The sequential ring feature enables you to create a list of up to five phone numbers. When the workspace
-        receives incoming calls, these numbers will ring one after another.
-        The sequential ring criteria specify settings such as schedule and incoming numbers for which to sequentially
-        ring or not.
+        Create Selective Reject Criteria Settings for an entity.
+
+        With the Selective Reject feature, you can reject calls at specific times from specific callers. This setting
+        takes precedence over Selectively Accept Calls.
+        Schedules can also be set up for this feature during certain times of the day or days of the week.
 
         **NOTE**: This API is only available for professional licensed workspaces.
 
         :param entity_id: Unique identifier for the entity.
         :type entity_id: str
         :param settings: new settings to be applied.
-        :type settings: SelectiveCriteria
+        :type settings: SelectiveRejectCriteria
         :param org_id: ID of the organization within which the entity resides. Only admin users of another
             organization (such as partners) may use this parameter as the default is the same organization as the
             token used to access API.
@@ -172,19 +154,19 @@ class SequentialRingApi(PersonSettingsApiChild):
         if org_id is not None:
             params['orgId'] = org_id
         body = settings.update()
-
-        url = self.f_ep(entity_id, f'criteria')
+        url = self.f_ep(entity_id, 'criteria')
         data = super().post(url, params=params, json=body)
         r = data['id']
         return r
 
     def read(self, entity_id: str,
-             org_id: str = None) -> SequentialRing:
+             org_id: str = None) -> SelectiveReject:
         """
-        Retrieve sequential ring settings for an entity.
+        Retrieve Selective Reject Settings for an entity.
 
-        The sequential ring feature enables you to create a list of up to five phone numbers. When the workspace
-        receives incoming calls, these numbers will ring one after another.
+        With the Selective Reject feature, you can reject calls at specific times from specific callers. This setting
+        takes precedence over Selectively Accept Calls.
+        Schedules can also be set up for this feature during certain times of the day or days of the week.
 
         **NOTE**: This API is only available for professional licensed workspaces.
 
@@ -194,30 +176,30 @@ class SequentialRingApi(PersonSettingsApiChild):
             organization (such as partners) may use this parameter as the default is the same organization as the
             token used to access API.
         :type org_id: str
-        :rtype: :class:`SequentialRing`
+        :rtype: :class:`SelectiveReject`
         """
         params = {}
         if org_id is not None:
             params['orgId'] = org_id
         url = self.f_ep(entity_id)
         data = super().get(url, params=params)
-        r = SequentialRing.model_validate(data)
+        r = SelectiveReject.model_validate(data)
         return r
 
-    def configure(self, entity_id: str, settings: SequentialRing,
-                  org_id: str = None):
+    def configure(self, entity_id: str, settings: SelectiveReject, org_id: str = None):
         """
-        Modify sequential ring settings for an entity.
+        Modify Selective Reject Settings for an entity.
 
-        The sequential ring feature enables you to create a list of up to five phone numbers. When the entity
-        receives incoming calls, these numbers will ring one after another.
+        With the Selective Reject feature, you can reject calls at specific times from specific callers. This setting
+        takes precedence over Selectively Accept Calls.
+        Schedules can also be set up for this feature during certain times of the day or days of the week.
 
         **NOTE**: This API is only available for professional licensed workspaces.
 
         :param entity_id: Unique identifier for the entity.
         :type entity_id: str
-        :param settings: Settings for new criteria
-        :type settings: SequentialRing
+        :param settings: new settings to be applied.
+        :type settings: SelectiveReject
         :param org_id: ID of the organization within which the entity resides. Only admin users of another
             organization (such as partners) may use this parameter as the default is the same organization as the
             token used to access API.