+++ wxc_sdk/workspace_settings/__init__.py
@@ -24,6 +24,7 @@ from ..person_settings.permissions_in import IncomingPermissionsApi
 from ..person_settings.permissions_out import OutgoingPermissionsApi
 from ..person_settings.privacy import PrivacyApi
 from ..person_settings.push_to_talk import PushToTalkApi
+from ..person_settings.selective_reject import SelectiveRejectApi
 from ..person_settings.sequential_ring import SequentialRingApi
 from ..person_settings.sim_ring import SimRingApi
 from ..person_settings.voicemail import VoicemailApi
@@ -59,6 +60,7 @@ class WorkspaceSettingsApi(ApiChild, base='workspaces'):
     permissions_out: OutgoingPermissionsApi
     privacy: PrivacyApi
     push_to_talk: PushToTalkApi
+    selective_reject: SelectiveRejectApi
     sequential_ring: SequentialRingApi
     sim_ring: SimRingApi
     voicemail: VoicemailApi
@@ -83,6 +85,7 @@ class WorkspaceSettingsApi(ApiChild, base='workspaces'):
         self.permissions_out = OutgoingPermissionsApi(session=session, selector=ApiSelector.workspace)
         self.privacy = PrivacyApi(session=session, selector=ApiSelector.workspace)
         self.push_to_talk = PushToTalkApi(session=session, selector=ApiSelector.workspace)
+        self.selective_reject = SelectiveRejectApi(session=session, selector=ApiSelector.workspace)
         self.sequential_ring = SequentialRingApi(session=session, selector=ApiSelector.workspace)
         self.sim_ring = SimRingApi(session=session, selector=ApiSelector.workspace)
         self.voicemail = VoicemailApi(session=session, selector=ApiSelector.workspace)